requirements.txt
streamlit==0.92.0
pandas==1.3.3
numpy==1.21.2
matplotlib==3.4.3
scikit-learn==0.24.2
sqlite3
psycopg2-binary
